Don't fail when source file is missing
A stacktrace might include relative paths to source files. When a code,
that had failed, comes from an external package that is built using a
native code (for example with Cython), calling `open` will raise the
"file not found" exception. As the result, sending a notification will
fail on creating payload and the exception won't be reported (except
the local logs if configured).

Leave `source` key empty when the file is not available.